# api/API/openapi/API_v2.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Kenku v2",
    "version": "2.0"
  },
  "paths": {
...
    "/v2/Settings/UserAgent": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get the current UserAgent used by Kenku",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Set a new UserAgent",
        "requestBody": {
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Settings"
        ],
        "summary": "Reset the UserAgent to default",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/v2/Settings/ImageCompressionLevel/{level}": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Set the Image-Compression-Level for Images",
        "parameters": [
          {
            "name": "level",
            "in": "path",
            "description": "100 to disable, 0-99 for JPEG ImageCompression-Level",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Level outside permitted range",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/Settings/BWImages/{enabled}": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Enable/Disable conversion of Images to Black and White",
        "parameters": [
          {
            "name": "enabled",
            "in": "path",
            "description": "true to enable",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v2/Settings/ChapterNamingScheme": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Gets the Chapter Naming Scheme",
        "description": "Placeholders:\n%M Obj Name\n%V Volume\n%C Chapter\n%T Title\n%A Author (first in list)\n%I Chapter Internal ID\n%i Obj Internal ID\n%Y Year (Obj)\n            \n?_(...) replace _ with a value from above:\nEverything inside the braces will only be added if the value of %_ is not null",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Sets the Chapter Naming Scheme",
        "description": "Placeholders:\n%M Obj Name\n%V Volume\n%C Chapter\n%T Title\n%A Author (first in list)\n%Y Year (Obj)\n            \n?_(...) replace _ with a value from above:\nEverything inside the braces will only be added if the value of %_ is not null",
        "requestBody": {
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v2/Settings/FlareSolverr/Url": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Sets the FlareSolverr-URL",
        "requestBody": {
          "description": "URL of FlareSolverr-Instance",
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Settings"
        ],
        "summary": "Resets the FlareSolverr-URL (HttpClient does not use FlareSolverr anymore)",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v2/Settings/FlareSolverr/Test": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Test FlareSolverr",
        "responses": {
          "200": {
            "description": "FlareSolverr is working!"
          },
          "500": {
            "description": "FlareSolverr is not working"
          }
        }
      }
    },
...
    "/v2/Settings/DownloadLanguage/{Language}": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Sets the language in which Series are downloaded",
        "parameters": [
          {
            "name": "Language",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v2/Settings/LibraryRefresh": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Sets the time when Libraries are refreshed",
        "requestBody": {
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/PatchLibraryRefreshRecord"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/PatchLibraryRefreshRecord"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/PatchLibraryRefreshRecord"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/PatchLibraryRefreshRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/v2/Settings/Metron": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "Sets Metron (metron.cloud) metadata credentials",
        "requestBody": {
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/SetMetronRecord"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/SetMetronRecord"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/SetMetronRecord"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/SetMetronRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Settings"
        ],
        "summary": "Clears Metron credentials (disables Metron lookups)",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/v2/Settings/DownloadClients": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Lists the configured download clients.",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DownloadClientConfig"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "Adds a download client.",
        "requestBody": {
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/SetDownloadClientRecord"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/SetDownloadClientRecord"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/SetDownloadClientRecord"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/SetDownloadClientRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadClientConfig"
                }
              }
            }
          },
          "400": {
            "description": "Name and base URL are required",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Updates an existing download client.",
        "requestBody": {
          "content": {
            "application/json-patch+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/SetDownloadClientRecord"
              }
            },
            "application/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/SetDownloadClientRecord"
              }
            },
            "text/json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/SetDownloadClientRecord"
              }
            },
            "application/*+json; x-version=2.0": {
              "schema": {
                "$ref": "#/components/schemas/SetDownloadClientRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Name and base URL are required",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No download client with that id",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v2/Settings/DownloadClients/{id}": {
      "delete": {
        "tags": [
          "Settings"
        ],
        "summary": "Removes a download client.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "No download client with that id",
            "content": {
              "text/plain; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json; x-version=2.0": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DownloadClientConfig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/DownloadClientType"
          },
          "baseUrl": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DownloadClientType": {
        "enum": [
          "QBittorrent"
        ],
        "type": "string"
      },
...
      "LibraryRefreshSetting": {
        "enum": [
          "AfterAllFinished",
          "AfterMangaFinished",
          "AfterEveryChapter",
          "WhileDownloading"
        ],
        "type": "string"
      },
...
      "PatchLibraryRefreshRecord": {
        "required": [
          "setting"
        ],
        "type": "object",
        "properties": {
          "setting": {
            "$ref": "#/components/schemas/LibraryRefreshSetting"
          },
          "refreshLibraryWhileDownloadingEveryMinutes": {
            "type": "integer",
            "description": "When API.Workers.LibraryRefreshSetting.WhileDownloading is selected, update the time between refreshes",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "SetDownloadClientRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/DownloadClientType"
          },
          "baseUrl": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SetMetronRecord": {
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "minLength": 1,
            "type": "string",
            "description": "Metron account username"
          },
          "password": {
            "minLength": 1,
            "type": "string",
            "description": "Metron account password"
          }
        },
        "additionalProperties": false,
        "description": "Metron (metron.cloud) account credentials for metadata lookups."
      },
...
    }
  },
...
}